标签页展示

+++ app/Http/Controllers/DingdanController.php
@@ -197,10 +197,6 @@ class DingdanController extends Controller
         }
         $order_bh = rand(100,999);
 
-        $coupon = Coupon::where('price',$req->yhj_1)->first();
-        $cu = Coupon_user::where('coupon_id',$coupon->id)->where('user_id',\Session::get('id'))->first();
-        $cu->delete();
-
         $dd = new Order;
         $dd -> zhuangtai = 2;
         $dd -> wuliu_id = $req -> wuliu_id;
@@ -311,10 +307,12 @@ class DingdanController extends Controller
             return back()->with('error','请选择收货地址');
         }
         $order_bh = rand(100,999);
-
-        $coupon = Coupon::where('price',$req->yhj_1)->first();
-        $cu = Coupon_user::where('coupon_id',$coupon->id)->where('user_id',\Session::get('id'))->first();
-        $cu->delete();
+        
+        if($req->yhj_1){
+            $coupon = Coupon::where('price',$req->yhj_1)->first();
+            $cu = Coupon_user::where('coupon_id',$coupon->id)->where('user_id',\Session::get('id'))->first();
+            $cu->delete();
+        }
 
         $dd = new Order;
         $dd -> zhuangtai = 3;
